Initial release (#1)
* Add: rollup, test

* Build, add demo page

* Fix: undefined error

* Update styles

* Style button

+++ src/index.js
@@ -1,7 +1,8 @@
 /* eslint-env greasemonkey */
-/* global $inline GM_addValueChangeListener */
+/* global $inline */
 const {createPref, createView} = require("webext-pref");
-const EventLite = require("event-lite");
+
+const createGMStorage = require("./storage");
 
 function GM_webextPref({
   default: _default,
@@ -13,6 +14,10 @@ function GM_webextPref({
   const initializing = pref.connect(createGMStorage());
   let isOpen = false;
   
+  if (typeof GM_registerMenuCommand === "function") {
+    GM_registerMenuCommand(`${getTitle()} - Configure`, openDialog);
+  }
+  
   return Object.assign(pref, {
     ready: () => initializing,
     openDialog
@@ -23,13 +28,18 @@ function GM_webextPref({
       return;
     }
     isOpen = true;
+    
+    let destroyView;
+    
     const modal = document.createElement("div");
     modal.className = "webext-pref-modal";
     modal.onclick = e => {
       if (e.target === modal) {
         modal.classList.remove("webext-pref-modal-open");
         modal.addEventListener("transitionend", () => {
-          destroyView(); // eslint-disable-line no-use-before-define
+          if (destroyView) {
+            destroyView();
+          }
           modal.remove();
           isOpen = false;
         });
@@ -51,87 +61,45 @@ function GM_webextPref({
           <style class="dialog-style"></style>
         </head>
         <body>
-          <h1 class="dialog-title"></h1>
           <div class="dialog-body"></div>
         </body>
       </html>
     `;
-    iframe.contentDocument.querySelector(".dialog-style").textContent = $inline("dialog.css|cssmin|stringify");
-    iframe.contentDocument.querySelector(".dialog-title").textContent = getTitle();
     
-    const destroyView = createView({
-      pref,
-      body,
-      translate,
-      root: iframe.contentDocument.querySelector(".dialog-body"),
-      getNewScope
-    });
-    
-    modal.appendChild(iframe);
+    modal.append(style, iframe);
     document.body.appendChild(modal);
     
-    // calc iframe size
-    iframe.style = `
-      width: ${iframe.contentDocument.body.offsetWidth}px;
-      height: ${iframe.contentDocument.body.scrollHeight}px;
-    `;
-    
-    modal.classList.add("webext-pref-modal-open");
+    iframe.onload = () => {
+      iframe.onload = null;
+      
+      iframe.contentDocument.querySelector(".dialog-style").textContent = $inline("dialog.css|cssmin|stringify");
+      
+      destroyView = createView({
+        pref,
+        body,
+        translate,
+        root: iframe.contentDocument.querySelector(".dialog-body"),
+        getNewScope
+      });
+      
+      const title = document.createElement("h2");
+      title.className = "dialog-title";
+      title.textContent = getTitle();
+      iframe.contentDocument.querySelector(".webext-pref-toolbar").prepend(title);
+      
+      // calc iframe size
+      iframe.style = `
+        width: ${iframe.contentDocument.body.offsetWidth}px;
+        height: ${iframe.contentDocument.body.scrollHeight}px;
+      `;
+      
+      modal.classList.add("webext-pref-modal-open");
+    };
   }
   
   function getTitle() {
     return typeof GM_info === "object" ?
       GM_info.script.name : GM.info.script.name;
-  }
-}
-
-function createGMStorage() {
-  const setValue = typeof GM_setValue === "function" ? promisify(GM_setValue) : GM.setValue.bind(GM);
-  const getValue = typeof GM_getValue === "function" ? promisify(GM_getValue) : GM.getValue.bind(GM);
-  const events = new EventLite;
-  
-  if (typeof GM_addValueChangeListener === "function") {
-    GM_addValueChangeListener("webext-pref-message", (name, oldValue, newValue) => {
-      events.emit("change", JSON.parse(newValue));
-    });
-  }
-  
-  return Object.assign(events, {getMany, setMany});
-  
-  function getMany(keys) {
-    return Promise.all(keys.map(k => 
-      getValue(`webext-pref/${k}`)
-        .then(value => [k, JSON.parse(value)])
-    ))
-      .then(entries => {
-        const output = {};
-        for (const [key, value] of entries) {
-          output[key] = value;
-        }
-        return output;
-      });
-  }
-  
-  function setMany(changes) {
-    return Promise.all(Object.entries(changes).map(([key, value]) => 
-      setValue(`webext-pref/${key}`, JSON.stringify(value))
-    ))
-      .then(() => {
-        if (typeof GM_addValueChangeListener === "function") {
-          return setValue("webext-pref-message", JSON.stringify(changes));
-        }
-        events.emit("change", changes);
-      });
-  }
-  
-  function promisify(fn) {
-    return (...args) => {
-      try {
-        return Promise.resolve(fn(...args));
-      } catch (err) {
-        return Promise.reject(err);
-      }
-    };
   }
 }
